requirements.txt
cryptography 
pycryptodome